+++ backend/app/api/git.py
@@ -2,7 +2,11 @@
 # visibility (require_app_member); webhook reads are scoped to the linked app's
 # visibility (an unlinked webhook stays panel-wide viewer-visible). Webhook
 # mutations and the Gitea host lifecycle stay admin-only.
-"""Git Server API endpoints for managing integrated Gitea instance and webhooks."""
+"""Git webhook + deployment API — the deploy-pipeline half.
+
+The Gitea self-host surface (lifecycle + repo browsing) moved to the
+serverkit-git extension (plan 52 Phase 6), which mounts it back at this same
+/api/v1/git prefix when installed."""
 
 from flask import Blueprint, request, jsonify
 
@@ -11,9 +15,7 @@ from ..middleware.rbac import (
     get_current_user, app_access_tier,
 )
 from ..models import Application, GitWebhook
-from ..services.git_service import GitService
 from ..services.webhook_service import WebhookService
-from ..services.gitea_api_service import GiteaAPIService
 from ..services.git_deploy_service import GitDeployService
 
 git_bp = Blueprint('git', __name__)
@@ -34,85 +36,6 @@ def _webhook_visible_to(user, webhook):
     return app_access_tier(user, app) is not None
 
 
-@git_bp.route('/status', methods=['GET'])
-@viewer_required
-def get_status():
-    """Get Gitea installation status."""
-    result = GitService.get_gitea_status()
-    return jsonify(result), 200
-
-
-@git_bp.route('/requirements', methods=['GET'])
-@viewer_required
-def get_requirements():
-    """Get resource requirements for Gitea installation."""
-    result = GitService.get_gitea_resource_requirements()
-    return jsonify(result), 200
-
-
-@git_bp.route('/install', methods=['POST'])
-@admin_required
-def install():
-    """Install Gitea with PostgreSQL."""
-    data = request.get_json() or {}
-
-    result = GitService.install_gitea(
-        admin_user=data.get('adminUser', 'admin'),
-        admin_email=data.get('adminEmail'),
-        admin_password=data.get('adminPassword')
-    )
-
-    if result.get('success'):
-        return jsonify(result), 201
-    return jsonify(result), 400
-
-
-@git_bp.route('/uninstall', methods=['POST'])
-@admin_required
-def uninstall():
-    """Uninstall Gitea and optionally remove data."""
-    data = request.get_json() or {}
-
-    result = GitService.uninstall_gitea(
-        remove_data=data.get('removeData', False)
-    )
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 400
-
-
-@git_bp.route('/start', methods=['POST'])
-@admin_required
-def start():
-    """Start Gitea server."""
-    result = GitService.start_gitea()
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 400
-
-
-@git_bp.route('/stop', methods=['POST'])
-@admin_required
-def stop():
-    """Stop Gitea server."""
-    result = GitService.stop_gitea()
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 400
-
-
-@git_bp.route('/restart', methods=['POST'])
-@admin_required
-def restart():
-    """Restart Gitea server."""
-    result = GitService.restart_gitea()
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 400
 
 
 # ==================== WEBHOOK ENDPOINTS ====================
@@ -270,163 +193,6 @@ def receive_webhook(token):
         payload=request.get_data(),
         payload_json=request.get_json(silent=True)
     )
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 400
-
-
-# ==================== REPOSITORY ENDPOINTS ====================
-
-@git_bp.route('/repos', methods=['GET'])
-@viewer_required
-def list_repositories():
-    """List all repositories in Gitea."""
-    token = request.args.get('token')
-    limit = request.args.get('limit', 50, type=int)
-
-    result = GiteaAPIService.list_repositories(token=token, limit=limit)
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 400
-
-
-@git_bp.route('/repos/<owner>/<repo>', methods=['GET'])
-@viewer_required
-def get_repository(owner, repo):
-    """Get repository details."""
-    token = request.args.get('token')
-
-    result = GiteaAPIService.get_repository(owner, repo, token=token)
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 404
-
-
-@git_bp.route('/repos/<owner>/<repo>/stats', methods=['GET'])
-@viewer_required
-def get_repo_stats(owner, repo):
-    """Get repository statistics."""
-    token = request.args.get('token')
-
-    result = GiteaAPIService.get_repo_stats(owner, repo, token=token)
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 400
-
-
-@git_bp.route('/repos/<owner>/<repo>/branches', methods=['GET'])
-@viewer_required
-def list_branches(owner, repo):
-    """List repository branches."""
-    token = request.args.get('token')
-
-    result = GiteaAPIService.list_branches(owner, repo, token=token)
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 400
-
-
-@git_bp.route('/repos/<owner>/<repo>/branches/<branch>', methods=['GET'])
-@viewer_required
-def get_branch(owner, repo, branch):
-    """Get branch details."""
-    token = request.args.get('token')
-
-    result = GiteaAPIService.get_branch(owner, repo, branch, token=token)
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 404
-
-
-@git_bp.route('/repos/<owner>/<repo>/commits', methods=['GET'])
-@viewer_required
-def list_commits(owner, repo):
-    """List repository commits."""
-    token = request.args.get('token')
-    branch = request.args.get('branch')
-    page = request.args.get('page', 1, type=int)
-    limit = request.args.get('limit', 30, type=int)
-
-    result = GiteaAPIService.list_commits(
-        owner, repo,
-        branch=branch,
-        page=page,
-        limit=limit,
-        token=token
-    )
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 400
-
-
-@git_bp.route('/repos/<owner>/<repo>/commits/<sha>', methods=['GET'])
-@viewer_required
-def get_commit(owner, repo, sha):
-    """Get commit details."""
-    token = request.args.get('token')
-
-    result = GiteaAPIService.get_commit(owner, repo, sha, token=token)
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 404
-
-
-@git_bp.route('/repos/<owner>/<repo>/contents', methods=['GET'])
-@viewer_required
-def list_files(owner, repo):
-    """List files in repository directory."""
-    token = request.args.get('token')
-    ref = request.args.get('ref', 'main')
-    path = request.args.get('path', '')
-
-    result = GiteaAPIService.list_files(owner, repo, ref=ref, path=path, token=token)
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 400
-
-
-@git_bp.route('/repos/<owner>/<repo>/contents/<path:filepath>', methods=['GET'])
-@viewer_required
-def get_file_content(owner, repo, filepath):
-    """Get file content."""
-    token = request.args.get('token')
-    ref = request.args.get('ref', 'main')
-
-    result = GiteaAPIService.get_file_content(owner, repo, filepath, ref=ref, token=token)
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 404
-
-
-@git_bp.route('/repos/<owner>/<repo>/readme', methods=['GET'])
-@viewer_required
-def get_readme(owner, repo):
-    """Get repository README."""
-    token = request.args.get('token')
-    ref = request.args.get('ref')
-
-    result = GiteaAPIService.get_readme(owner, repo, ref=ref, token=token)
-
-    if result.get('success'):
-        return jsonify(result), 200
-    return jsonify(result), 404
-
-
-@git_bp.route('/version', methods=['GET'])
-@viewer_required
-def get_gitea_version():
-    """Get Gitea server version."""
-    result = GiteaAPIService.get_server_version()
 
     if result.get('success'):
         return jsonify(result), 200
